pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = "Order management backend."
authors = ["codecakes <akulmat@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
ujson = "^5.4.0"
pydantic = "~=1.10.2"
Cython = "^0.29.27"
requests = "~=2.28.1"
Authlib = "~=1.1.0"
markdown = "~=3.4.1"
django = "~=4.1.3"
django-filter = "~=22.1"
djangorestframework = "~=3.14.0"
types-requests = "~2.28.8"
gunicorn = "20.1.0"
hypercorn = "0.14.3"

[tool.poetry.dev-dependencies]
pytest = "7.2.0"
black = {extras = ["d", "click>=8.0.2"], version = ">=22.10.0"}
pre-commit = "^2.17.0"
pylint = "^2.15.5"
isort = "^5.10.1"
anyio = "~3.6.2"
pytest-asyncio = "0.20.1"
pysqlite3 = "^0.4.6"
types-pytest-lazy-fixture = "^0.6.3.1"
mypy = "^0.982"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "fast: fast executing tests. use this marker if unsure."
]
asyncio_mode = "auto"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
addopts = ["-ra -q", "--color=yes",]

# See: https://github.com/python/mypy/issues/5205
# mypy global options:
[tool.mypy]
python_version = "^3.10"
plugins = ["pydantic.mypy"]
# See: https://blog.wolt.com/engineering/2021/09/30/professional-grade-mypy-configuration/
disallow_untyped_defs = true
check_untyped_defs = true
disallow_any_unimported = true
ignore_missing_imports = false
no_implicit_optional = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = true

# See: https://pydantic-docs.helpmanual.io/mypy_plugin/
# mypy per-module options:
[tool.pydantic-mypy]
init_typed = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
